+++ gui/Rectangle.java
@@ -9,13 +9,12 @@ import java.awt.Color;
 public class Rectangle {
 	
 	public final double x;
-//	public final int x;
-	public final int y;
-	public final int width;
-	public final int height;
+	public final double y;
+	public final double width;
+	public final double height;
 	public final Color color;
 
-	public Rectangle(int x1, int y1, int x2, int y2, Color color) {
+	public Rectangle(double x1, double y1, double x2, double y2, Color color) {
 		this.x = x1;
 		this.y = y1;
 		this.width = x2;
